+++ src/mimic/orm/macros/src/node/mod.rs
@@ -82,7 +82,7 @@ pub trait MacroNode: Schemable {
     // formats the code needed to send something via ctor to the schema
     #[must_use]
     fn ctor_schema(&self) -> TokenStream {
-        let ctor_fn = format_ident!("ctor_{}", lib_rand::next_u64());
+        let ctor_fn = format_ident!("ctor_{}", ic::rand::next_u64());
         let schema = self.schema();
 
         quote! {